refactor(providers): extract decodeEntities into a shared module
Addresses CodeRabbit's finding on #1639: decodeEntities()/NAMED_ENTITIES was
duplicated verbatim across successfactors.mjs, dassault.mjs, and
softgarden.mjs — the exact drift that let the range guard go stale in the
first place. Moved it to providers/_html-entities.mjs (with the fixed guard)
and pointed all three at it.

rheinmetall.mjs carried the same duplicated (and still-buggy) copy; folded it
into the same extraction and added the matching regression test.
