+++ providers/dassault.mjs
@@ -1,5 +1,6 @@
 // @ts-check
 /** @typedef {import('./_types.js').Provider} Provider */
+import { decodeEntities } from './_html-entities.mjs';
 
 // Dassault Systèmes provider — hits the public Exalead "card search" API that
 // powers www.3ds.com/careers/jobs. Single-company provider (like ibm.mjs): the
@@ -42,24 +43,6 @@ export function buildUrl(start) {
   for (const r of REFINES) p.append('r', r);
   p.set('start', String(start));
   return `${BASE}?${p.toString()}`;
-}
-
-// Minimal HTML entity decoder — titles/URLs/locations carry named (&amp;, &apos;)
-// and numeric (&#252; / &#xfc;) entities. Mirrors successfactors.mjs.
-const NAMED_ENTITIES = { amp: '&', lt: '<', gt: '>', quot: '"', apos: "'", nbsp: ' ' };
-/** @param {string} s */
-function decodeEntities(s) {
-  return s.replace(/&(#x?[0-9a-fA-F]+|[a-zA-Z]+);/g, (m, body) => {
-    if (body[0] === '#') {
-      const code = body[1] === 'x' || body[1] === 'X' ? parseInt(body.slice(2), 16) : parseInt(body.slice(1), 10);
-      // String.fromCodePoint throws RangeError outside 0..0x10FFFF or on a lone
-      // surrogate half — a malformed/adversarial entity must degrade to the
-      // original text, never crash the whole parse.
-      const valid = Number.isFinite(code) && code >= 0 && code <= 0x10ffff && !(code >= 0xd800 && code <= 0xdfff);
-      return valid ? String.fromCodePoint(code) : m;
-    }
-    return NAMED_ENTITIES[body.toLowerCase()] ?? m;
-  });
 }
 
 // Pull every <Meta name="X"><MetaString name="value">V</MetaString> pair from one
